适配新 IO_Board 板

+++ inc/Switch_Boards/Switch_Board/GPIO.hpp
@@ -61,7 +61,7 @@ class C_GPIO : private Robot_Hardware
 public:
     int Get_Yaml_And_Init(const shared_ptr<Device_class>& Device, const YAML::Node& One_Node);
 
-    void Get_Buttons_Value(const shared_ptr<Device_class> &Device, int *Res);
+    void Get_Buttons_Value(const shared_ptr<Device_class>& Device, u8* Res);
 
     void Set_GPIOx_To_Input_Mode(const shared_ptr<Device_class> &Device, GPIO_TypeDef *GPIOx, u32 GPIOx_Pin);
     u8 Get_GPIOx_Value(const GPIO_TypeDef *GPIOx, u32 GPIOx_Pin);
